src/content/image-processing-service/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pillow==11.0.0
pydantic==2.10.3
pydantic-settings==2.6.1
python-multipart==0.0.12
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1
